refactor: upgrade module
